requirements.txt
a2a-sdk>=0.1.0
httpx>=0.24.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-adk
litellm
